// apps/web/types/Product.ts
export type Product = {
    id:string | number
    productTitle: string;
    productDescription: string;
    category: string;
    brand: string;
    price: string;
    discountPercentage: string;
    stockStatus: string;
    minimumOrderQuantity: number;
    warranty: string;
    shippingInfo: string;
    returnPolicy: string;
  };